refactor: unify claude/codex/opencode/cursor behind a Harness trait + registry
Collapse the three parallel per-harness surfaces (detection in
harness.rs, chat run_turn in chat/*, skill install in install_skills.rs)
into one Harness trait with per-harness impls under src/local/harness/.

- Detection, chat-turn, and skill-install are now optional capabilities
  on one trait; each consumer iterates a single registry() and filters
  by capability (supports_chat / skill_target).
- Removes the stringly-typed dispatch match and the triple-duplicated
  harness id lists (HARNESS_IDS, detect assembly, install Agent enum).
- Cursor becomes a first-class install-only harness impl.
- Adding a harness is one new file + one registry() line.

No behavior change: detection order, wire output, and install paths are
preserved.
